refactor: relocate GA sdk
